Location Finder App
This app finds the correct location based on the input address whether it's properly formatted or not. Moreover it also gives the street details, region details, weather conditions of the Location.

User just simply need to either enter the address or by entering the latitude and longitude. It will give you the location and display it on Map.